Update the README with new files names and rename a couple more
